requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
